Moved dynamic creation of Services to ServiceCollection class. Updated Client and Account to extend it.
